requirements.txt
streamlit
google.generativeai
pypdf
python-docx 
opencv-python
pillow  
